backend/requirements.txt
bcrypt>=4.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
openai>=1.50.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
httpx>=0.27.0
qdrant-client>=1.7.0
openpyxl>=3.1.0
reportlab>=4.0.0
python-pptx>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
markdown>=3.5.0
Pillow>=10.0.0
apscheduler>=3.10.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0
redis>=5.0.0
matplotlib>=3.8.0
numpy>=1.26.0
jsonschema>=4.20.0
beautifulsoup4>=4.12.0
exa-py>=1.0.0
youtube-transcript-api>=1.0.0
deep-translator>=1.11.0
